Issue #3991: Split and Organize Checkstyle inputs by Test for ExecutableStatementCount
