+++ src/test/java/com/puppycrawl/tools/checkstyle/checks/sizes/ExecutableStatementCountCheckTest.java
@@ -40,7 +40,8 @@ public class ExecutableStatementCountCheckTest
     @Override
     protected String getPath(String filename) throws IOException {
         return super.getPath("checks" + File.separator
-                + "sizes" + File.separator + filename);
+                + "sizes" + File.separator + "executablestatementcount"
+                + File.separator + filename);
     }
 
     @Test
